+++ enocean_async/protocol/erp1/telegram.py
@@ -1,10 +1,7 @@
 from dataclasses import dataclass
 from enum import IntEnum
 
-from enocean_async.eep.manufacturer import Manufacturer
-
 from ...address import EURID, Address, BaseAddress, BroadcastAddress
-from ...eep.id import EEP
 from ..esp3.packet import ESP3Packet, ESP3PacketType
 from .errors import ERP1ParseError
 from .rorg import RORG
@@ -308,37 +305,3 @@ class ERP1Telegram:
     def repeater_count(self) -> RepeaterCount:
         """Return lower 4 bits (last 4 bits) of the 8-bit status byte, interpreted as per ERP1 spec as repeater count."""
         return RepeaterCount(self.status & 0x0F)
-
-
-class FourBSTeachInVariation(IntEnum):
-    UnidirectionalProfileLess = 0
-    UnidirectionalWithProfileOrBidirectional = 1
-
-
-@dataclass
-class FourBSTeachInTelegram:
-    teach_in_variation: FourBSTeachInVariation = 1
-    eep: EEP | None = None
-
-    @classmethod
-    def from_erp1(cls, erp1: ERP1Telegram) -> "FourBSTeachInTelegram":
-        if erp1.rorg != RORG.RORG_4BS:
-            raise ValueError("Wrong RORG.")
-
-        if not erp1.is_learning_telegram:
-            raise ValueError("Not a learning telegram")
-
-        teach_in_variation = FourBSTeachInVariation(erp1.bitstring_raw_value(24, 1))
-        eep: EEP | None = None
-
-        if (
-            teach_in_variation
-            == FourBSTeachInVariation.UnidirectionalWithProfileOrBidirectional
-        ):
-            func = erp1.bitstring_raw_value(0, 6)
-            type_ = erp1.bitstring_raw_value(6, 7)
-            manufacturer_id = erp1.bitstring_raw_value(13, 11)
-            manufacturer = Manufacturer(manufacturer_id)
-            eep = EEP(0xA5, func, type_, manufacturer)
-
-        return cls(teach_in_variation=teach_in_variation, eep=eep)